feat: allow Aranet4Advertisement to read advertisement data
This refactors out the actual data handling to the Aranet4Advertisement
class so clients can reuse this logic without completely using the
Aranet4Scanner class. (This will be useful in wrapping this library
where scanning is provided elsewhere, like in Home Assistant.)

Also removes unused has_service variable that was not accesisble.